+++ src/task/assoc/createassociationsjob.h
@@ -59,8 +59,8 @@ protected:
     //dbo name->ds_id->trs
 
     std::map<std::string,
-        std::map<unsigned int,
-            std::tuple<unsigned int, std::vector<std::pair<std::string, unsigned int>>>>> associations_;
+        std::map<unsigned long,
+            std::tuple<unsigned int, std::vector<std::pair<std::string, unsigned long>>>>> associations_;
     // dbcontent -> rec_num -> <utn, src rec_nums (dbcontent, rec_num)>
 
     std::map<std::string, std::pair<unsigned int,unsigned int>> association_counts_; // dbcontent -> total, assoc cnt
